Added -pthread flag into definitions.
Because -pthread is required to be passed to the compiler according to the Linux
Programmer's Manual (man 7 pthreads). We currently rely on CMake passing
arguments of ADD_DEFINITIONS() that don't look like preprocessor definitions
(e.g. start with -D or /D) to the compiler. This is likely not future-proof, but
will do for now. This does not appear to have an effect anyway.